+++ src/main/java/newfarma/unidadmedida/UnidadMedidaRepositoryJPA.java
@@ -3,7 +3,7 @@ package newfarma.unidadmedida;
 import newfarma.model.UnidadMedida;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnidadMedidaRepositoryJPA extends JpaRepository<UnidadMedida,Long> {
+public interface UnidadMedidaRepositoryJPA extends JpaRepository<UnidadMedida,Integer> {
     //@Query(value = "from UnidadMedida p where p.idunidadmedida = ?1 ")
     public boolean existsUnidadMedidaByNombreunidad(String nombre);
 }
